Refactor to single page architecture
Major Changes:
- Replace PortfolioLayout (850 lines) with PortfolioLayoutSimple (220 lines)
  - 74% code reduction
  - Remove height measurement logic, ResizeObserver, useLayoutEffect
  - Remove contentPaddingTop calculations and fade animations
  - Simplify to natural vertical flow (desktop) and sticky positioning (mobile)

- Update URL structure
  - Old: /works/[id]?keywordId=xxx
  - New: /?keywordId=xxx&workId=123
  - Add legacy URL redirect support

- Remove legacy components
  - CategorySidebar → StaticCategorySidebar (natural flow for desktop)
  - MobileCategoryMenu → MobileSwipeableCategories (with slider toggle)
  - WorkListScroller → WorkListScrollerFlex
  - Remove WorkSelectionProvider (URL-based work selection)

- Mobile improvements
  - Add MobileSwipeableCategories with horizontal swipe gestures
  - Add MobileCategorySlider with click-to-toggle functionality
  - Fix sticky positioning on mobile (hydration-safe mounting)
  - Fix overflow issues to show only one category view at a time

- Fix hydration errors
  - Add mounted state for hydration safety
  - Ensure server/client rendering consistency

- Clean up exports and documentation
  - Update component index exports
  - Remove legacy component references from documentation
  - Update LayoutStabilityContext comments
