+++ front/src/presentation/index.ts
@@ -2,7 +2,7 @@
 export * from './components';
 
 // Explicitly export layout components to ensure proper TypeScript resolution
-export { default as PortfolioLayout } from './components/layout/PortfolioLayout';
+export { default as PortfolioLayoutSimple } from './components/layout/PortfolioLayoutSimple';
 
 // Re-export UI primitives
 export * from './ui';
